+++ website/src/components/header/header.js
@@ -1,6 +1,10 @@
 import React from 'react';
 import logo from './sprite2.png';
-import Icons from './icons'
+import Icons from './Icons/Icons';
+import Home from './Buttons/Home_button';
+import About from './Buttons/About_button';
+import Work from './Buttons/Work_button';
+import Contact from './Buttons/Contact_button';
 import { BrowserRouter as Router, Route, Link } from "react-router-dom";
 
 
@@ -23,17 +27,11 @@ function Header() {
 };
 
 
-function chef(e) {
-  e.preventDefault();
-  console.log('The link was clicked.');
-}
-
-
 
 const posts = [
-  <li><button onClick={chef} > Home </button></li>,
-  <li><button onClick={chef} > About </button></li>,
-  <li><button onClick={chef} > Resume </button></li>,
-  <li><button onClick={chef} > Contact </button></li>
+  <li><Home/></li>,
+  <li><About/></li>,
+  <li><Work/></li>,
+  <li><Contact/></li>
 ];
 export default Header;
