--- website/src/components/header/props.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-
-export default function Blog(props) extends Component {
-  render() {
-    return (
-      <h1> This is MyComponent </h1>
-    );
-  }
-}
