+++ website/src/components/landing/landing.js
@@ -1,5 +1,5 @@
 import Header from '../header/header'
-import Home from '../pages/first'
+import Home from '../pages/Home/Home'
 import React, { Component } from 'react'
 
 class Land extends Component {
