--- website/src/components/pages/home.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Home(){
-  return (
-    <div className = "main_page">
-      <p> Tell me about yourself</p>
-    </div>
-  );
-}
-
-export default Home;
